requirements.txt
# VoteSocio — Dependencies
# Pipeline
duckdb>=1.1.0
pandas>=2.1.0
pyarrow>=14.0.0
pyyaml>=6.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.0
shapely>=2.0.0
pyproj>=3.6.0

# Analysis
scikit-learn>=1.4.0
statsmodels>=0.14.0
scipy>=1.12.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0

# Dev
ruff>=0.4.0
jupyter>=1.0.0
ipykernel>=6.0.0
